Из файла record.py вынесены в отдельные файлы SubField (subfield.py) и Field (field.py). Конструктор класса Field теперь поддерживает в качестве аргумента value: список объектов класса SubField, словарь из ключей и значений для инициализации объектов SubField,  а так же для задания атрибута value (если в словаре есть ключ в виде пустой строки).
